requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
django-cors-headers
django-debug-toolbar==4.1.0
django-dotenv==1.4.2
django-filter==23.2
django-import-export==3.2.0
djangorestframework
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
Markdown==3.4.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
psycopg2-binary
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
